+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.2] - 2026-Apr-19
+- dst_ips & dst_lists functionality
+- src_port functionality
+
 ## [0.2.1] - 2026-Apr-15
 - nftables counter write functionality
 - counters cli argument/opton